+++ types/src/block/output/nft.rs
@@ -316,6 +316,12 @@ impl NftOutput {
         &self.nft_id
     }
 
+    /// Returns the nft ID if not null, or creates it from the output ID.
+    #[inline(always)]
+    pub fn nft_id_non_null(&self, output_id: &OutputId) -> NftId {
+        self.nft_id.or_from_output_id(output_id)
+    }
+
     ///
     #[inline(always)]
     pub fn unlock_conditions(&self) -> &UnlockConditions {
@@ -352,7 +358,7 @@ impl NftOutput {
 
     /// Returns the nft address for this output.
     pub fn nft_address(&self, output_id: &OutputId) -> NftAddress {
-        NftAddress::new(self.nft_id().or_from_output_id(output_id))
+        NftAddress::new(self.nft_id_non_null(output_id))
     }
 
     ///